requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.1
